Ajustando equacao do sensor de PH

+++ ESP8266MQTT/Core/Inc/util/stm32_util.h
@@ -1,11 +1,17 @@
 #ifndef STM32_UTIL_H
 #define STM32_UTIL_H
 
+#include <stdbool.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <stdatomic.h>
+#include <math.h>
+#include "main.h"
 #include "stm32l4xx_hal.h"
-#include "stm32l4xx_hal_adc.h"
+#include "stm32l433xx.h"
 #include <util/dwt_delay.h>
 
-uint32_t stm32_util_read_analog(uint32_t channel);
+uint16_t stm32_util_read_analog(uint32_t channel);
 
 void stm32_util_set_pin_output(GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin);
